feat: added transfer history  feature without transastions (#12)
* feat: added transfer history  feature without transastions

* feat: corrected comments and refactored generator of xlsx transfer history

* feat: corrected xlsx generator of transfer history

* feat: corrected comments

* feat: refactored xlsx generator

+++ src/main/java/com/example/bank/client/jpa/Client.java
@@ -1,7 +1,7 @@
 package com.example.bank.client.jpa;
 
 import com.example.bank.auditing.AuditorEntity;
-import com.example.bank.bankTransfer.account.Account;
+import com.example.bank.bank.transfer.account.Account;
 import com.example.bank.client.Role;
 import com.example.bank.security.token.Token;
 import jakarta.persistence.*;
@@ -29,7 +29,7 @@ public class Client extends AuditorEntity implements UserDetails {
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "client_sequence")
     @Column(nullable = false)
     @EqualsAndHashCode.Exclude
-    private Integer id;
+    private Long id;
     @NotEmpty
     @Column(name = "first_name", nullable = false)
     private String firstName;
